removed unwanted files

+++ head_pose_estimation.py
@@ -30,7 +30,7 @@ def get_2d_points(img, rotation_vector, translation_vector, camera_matrix, val):
     point_3d.append((front_size, front_size, front_depth))
     point_3d.append((front_size, -front_size, front_depth))
     point_3d.append((-front_size, -front_size, front_depth))
-    point_3d = np.array(point_3d, dtype=np.float).reshape(-1, 3)
+    point_3d = np.array(point_3d, dtype=np.cfloat).reshape(-1, 3)
     
     # Map to 2d img points
     (point_2d, _) = cv2.projectPoints(point_3d,
